sys-devel/gcc-config: Renamed gcc-config-1.8-r00.2 to gcc-config-1.8-r1.
gcc-apple-4.2.1_p5666-r2.ebuild in gx86 needs to depend on this version of gcc-config, but -r00.2 is not a valid version in gx86. Therefore, I claim -r1 for this version, in conflict with the usual interrevision policy.

This is in conflict with the usual interrevision policy.

(Portage version: 2.2.14-prefix/hg/Linux x86_64, signed Manifest commit with key 0x10FB016B)