sys-libs/ncurses: fix build during bootstrap
When no ncurses is installed and not doing cross compilation, as during
bootstrap, we should not attempt to build cross tools, because configure
wasn't run to create them.
Normally this doesn't happen, because ncurses would be installed
already, but during a bootstrap this obviously isn't the case.

Package-Manager: Portage-2.3.18-prefix, Repoman-2.3.6