scripts/bootstrap-prefix: make FSF gcc the default compiler on ppc-macos
We need (or want) more recent features and standards, and FSF gcc-7
works fine, so use that instead of Clang (which we can't get a recent
version to compile).  Because of sys-libs/csu doing some weird stuff, we
really need gcc-apple for that, but that package got hacked to use it,
the rest is all fine with gcc-7.3.0-r3.