scripts/bootstrap-prefix.sh: allow direct rsync in bootstrap_tree
To test bootstrapping on a more recent tree than the latest snapshot,
allow special TREE_FROM_SRC envvar to specify the source to rsync after
unpacking a snapshot.  This is intended for bootstrap testing only,
normal usage should never use this, hence the support being a bit
brittle.  The exit code of this additional rsync step is ignored.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>