scripts/bootstrap-prefix: ensure USE is picked up in stage3
Portage seems to ignore USE= environment for build dependencies in a
different location.  Since we need this to be respected in order to
avoid many cycles, temporarily encode the requested USE-flags in
use.mask and use.force.  We use the profiles/ location which should get
wiped out hereafter, but also remove the files used whilst emerging to
reduce the risk of leaving this behind in someone's setup.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>