scripts/bootstrap-prefix: allow resume of stage3 after python-exec
Portage installs python-exec also in ROOT/tmp, which foobars the
manually bootstrapped python installation there, rendering it broken.
In a successful bootstrap this goes unnoticed, because the running
Portage (=Python) session installs Python later in the same merge
sequence.  As such, after that emerge, the python from ROOT is picked
up.  If we failed on one of the packages between python-exec and python,
however (~60 pkgs inbetween, including a compiler) we leave the setup in
a broken state.

Hence, prior to starting stage3, re-new the Python 2.7 symlinks to deal
with the scenario of resuming a broken off stage3.  Since ROOT/tmp gets
removed shortly after Python gets installed in ROOT, it is fine that we
effectively kill python-exec in ROOT/tmp here.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>