dev-lang/python: sync 3.6.6
Disabling ABIFLAGS is long gone in gentoo repo, to align with upstream
and other distros.  While --without-pymalloc would omit the 'm' abi
flag, this breaks 'import ctypes' on Cygwin.  However, bug#452720 is
handled differently now, so this is not a problem any more.
Not changing 3.6.5 in hope for less breakage, yet that one's patches
still apply to 3.6.6.
Bug: https://bugs.gentoo.org/452720
Package-Manager: Portage-2.3.55.1-prefix, Repoman-2.3.12

Signed-off-by: Michael Haubenwallner <haubi@gentoo.org>