scripts/bootstrap-prefix: unify libffi behaviour in bootstrap_gnu
due to commit 94a0ab4f, which adds -L${ROOT}/tmp/usr/lib64 to LDFLAGS
in bootstrap_python(), the python bootstrap build sees libffi.so and
succeeds.  but there is no -Wl,-rpath,${ROOT}/tmp/usr/lib64, and the
built _ctypes.so is unusable as the run time loader does not find
libffi.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>