scripts/bootstrap-prefix: fix make bootstrap
Due to wrongly patched make, it would segfault when globs are used (as
for m4-1.4.18 and sed-4.7).  This crash was previously ignored by
bootstrapping m4-1.4.17, but resurfaced in stage2 when sed-4.7 is
installed.

Just define __alloca and __stat for GNU/Linux hosts, because that should
work everywhere.  We technically don't need the m4 fix any more because
we can use (fixed) 1.4.18, but keep it around as it doesn't hurt much.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>