bootstrap-prefix.sh: fix circular deps in stage3
In prefix guest, when about to "with_stack_emerge_pkgs sys-apps/grep",
portage fails with circular dependencies potentially breakable with
USE=-berkdb, although we do set the USE=-berkdb environment variable.
The problem here is that USE=-berkdb does apply to the stacked prefix
only, not the base prefix, so we need to persist the base prefix' USE
flags when we are about to do "with_stack_emerge_pkgs", even if this is
not a problem with RAP right now, maybe because of a different tree.

Signed-off-by: Michael Haubenwallner <haubi@gentoo.org>