scripts/bootstrap-prefix: ensure stage1 gcc wrapper is functional
In the these days quite uncommon case of calling the stage functions
directly, stage1's gcc wrapper would use the unset variable STAGE1_PATH.
This variable is normally set in the interactive proze that re-defines
PATH to remove any unwanted stuff.

Avoid this dependency by storing the original value of PATH right at the
startup of the script, such that stage1 can refer to it, whether or not
called from the interactive wrapper or not.

Thanks Sargun Dhillon in bug #779145.

Closes: https://bugs.gentoo.org/779145
Signed-off-by: Fabian Groffen <grobian@gentoo.org>