bootstrap-prefix.sh: fix RAP bootstrap with non-local user, bug 766417
When the bootstrapping user is handled by Kerberos and is not local to
the host, the libc from prefix doesn't know how to convert usernames
to user ids, which ends up confusing portage. Using getent from the host
to add the bootstrapping user to ${EPREFIX}/etc/passwd works around this
issue.

Closes: https://bugs.gentoo.org/766417
Signed-off-by: Guilherme Amadio <amadio@gentoo.org>