Revert "scripts/bootstrap-prefix.sh: don't unset *FLAGS for stage3 @system"
Need to discuss this more with grobian, but I'd misunderstood
where we do this (it's _before_ the -e @system, not _after_,
so it's not like it's just before the user might update
their system anyway).

coreutils (and possibly libtasn1) are still broken but
I'm not convinced this is the right fix after grobian@
raised concerns.

Will revisit shortly but reverting for now to avoid
problems.

This reverts commit a3175e08822394188b112dc5e23475e876cd603e.

Bug: https://github.com/gentoo/prefix/pull/5
Signed-off-by: Sam James <sam@gentoo.org>