Revert "sys-devel/gcc-12.2.0-r1: revbump to drop nopie patch for Apple SI"
This reverts commit 32708ce1cab1fc98af5eccc8d2f39df761d528bf.

The original commit is wrong as a result of misunderstanding the bug
report. First, this fix should be applied to Intel, not ARM (and
preferably both), but the commit did exactly the opposite. Next, no
revbump is necessary since this bug / patch has no effect on Apple
Silicon, and on Intel it failed to build to begin with.

Bug: 898610
Signed-off-by: Yifeng Li <tomli@tomli.me>