scripts/bootstrap-prefix: keep LDFLAGS while emerging @system
revert a bit gone in through 14fbc8f14c4c484e1809e001e262efca9b6f6ecd

The final @system emerge of stage3 is pulling in the linker, so as long
as we don't have one installed we need manual LDFLAGS in order for the
host linker to find the libs from the new Prefix.

Problem observed in auto-bootstrap x86_64-apple-darwin17/20230710.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>