scripts: make bootstrap snapshot available on rsync slaves
In bug #913415 a bunch of confusing and wrong things came out, and in an
attempt to fix this, the historical "prefix-overlay" fetched from dev
was replaced with ordinary "portage" from rsync snapshots dir.  However,
this relies on the prefix mirror to provide the file, which can easily
be overridden by the user.
So retrieve the snapshot in use and retain it on the rsync slaves to
make this in the very least less confusing for those following the
locations, and ensure it works for people with custom mirror setups.

Bug: https://bugs.gentoo.org/913415
Signed-off-by: Fabian Groffen <grobian@gentoo.org>