scripts/bootstrap-prefix: fix RAP stage3 on python due to init.d script
Portage enforces init.d scripts must have a valid shebang, which means
openrc-run must exist on the system.  For RAP-systems this script is to
be provided by openrc (in contrast to baselayout for non-RAP Prefix
systems), however during bootstrap openrc is not pulled in, and in
reality it never is.  So provide a dummy openrc-run in the RAP path for
stage3 before python or rsync are pulled in, such that Portage's QA
check is happy about their init.d scripts.

Closes: https://bugs.gentoo.org/913856
Signed-off-by: Fabian Groffen <grobian@gentoo.org>