sys-devel/binutils-config: make ld arguments depending on OS release
This is somewhat shady, because we'd need the target ld's version, but
it's better to use the current platform version rather than the
requested platform version since the latter has nothing to do with which
linker is being used.

Not bumping the version, since existing installs don't have this issue,
and only new installs on Sanoma suffer from this, which never managed to
succeed, so basically don't exist yet.

Bug: https://bugs.gentoo.org/916291
Signed-off-by: Fabian Groffen <grobian@gentoo.org>