scripts/bootstrap-prefix: find OS matching SDK on Darwin
For some reason Apple installs SDKs for the next versions of the OS
automatically, and then enables those SDKs.  Doing so, however results
in mismatches between the running system and the SDK, such as finding
definitions of symbols, but not being able to resolve them during
runtime.

Try and find an SDK that matches the running OS, in the hope that things
match better.

Bug: https://bugs.gentoo.org/910720
Signed-off-by: Fabian Groffen <grobian@gentoo.org>