scripts/bootstrap-prefix: unbreak arm64-macos bootstrap
Trim down the guessing and more in configure_toolchain.

Problem is that configure_toolchain is called in all three stages, and
its output is used.  Since we need to bootstrap a compiler very early,
we need to rely on PATH for a compiler, which means in stage3 we always
pick up the stage2 emerged compiler.  This changes the output of the
function due to a change in env.

Avoid this by simply hardwiring the exceptional case, which is already
hardwired before already.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>