scripts/bootstrap-prefix: avoid final emerge -e @system
Long time ago, when we started bootstraps, all stages were done straight
into the destination Prefix, which meant tools could link to the host
and/or configured in an odd way where packages would refer to (config)
files from the host.  To make sure all of that was set straight, the
blunt tool emerge -e @system was run as last step to make sure
everything came from Portage proper.

Today, we bootstrap in early stages in EPREFIX/tmp and even set up a
Portage instance there that will use tools that may or may not come from
host or be configured to use it.  However, it cleanly pulls itself out
of the mud by letting Portage install @system in the destination Prefix,
thus it compiles every package it needs with the requirements as set out
by the ebuilds.

Thus there is no need to really recompile everything all over again.  It
should be fine with a simple recompile of all packages for which
USE-flags changed after we lifted some overrides to avoid circular or
unnecessary deps.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>