Revert "scripts/b-p.sh: link /usr/{lib -> lib64} in stage1"
This reverts commit 15ae1999d07446d243021983ee77d89f2335f141 as it
prevents using the prefix shortcut installation, i.e., when the host
is also a Gentoo installation *and* when lib64/ is not a symlink to
lib/ (which is quite common these days).

Furthermore, this check seems no longer necessary, as I was able to
bootstrap prefix on a Gentoo installation where lib64/ is *not* a
symlink to lib/ just fine.

Closes: https://bugs.gentoo.org/927957
Signed-off-by: Florian Schmaus <flow@gentoo.org>