sys-devel/binutils-config: remove duplicate -L and -rpath entries
Ever since CLT15.3 duplicate -L or -rpath entries result in warnings.
These warnings, albeit just ugly, make that some packages croak or
worse: abort, claiming the linker is broken (such as Ruby).  This is
nonsense, but cater for this by removing any duplicate paths, and not
adding any paths that are already added.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>