bootstrap-prefix: use full path for `do_emerge_pkgs`
In the stage3 bootstrap we always prefer to use tools that have been built for
stage3; to accomplish this we ensure that it is the first thing evaluated in PATH.
Unfortunately, Portage, Python, and Python-exec are often pulled into the depgraph
at some point before we're fully boostrapped. To ensure that we don't try and execute
${EPREFIX}/usr/bin/emerge before we're ready, always provide the full path to the
bootstrap Python interpreter and emerge script.

Signed-off-by: Matt Jolly <kangie@gentoo.org>