scripts/bootstrap-prefix: use GCC-5 on all Darwin bootstraps
With recent macOS we need GCC-14 in order to make sense of system
headers.  However, GCC-14 gets confused by (older) Clang installed on
the host system.  So we need GCC-5 to get us a C11/C++11 compiler that
can compile GCC-14 to do the real work.

Unfortunately for GCC-5 to work properly on systems that do not have
/usr/{include,lib} we need wrappers such as we use in Gentoo Prefix
proper, thus part of that needs to be pulled and put in place too in
order to get correctly compiled and running binaries.

This change succeeds bootstrap on Darwin 17, 19 and 20, 21 and 22
pending.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>