bootstrap-prefix: fix portability in "cp" usage
This script installs Prefix when you don't yet have it. It shouldn't
assume Gentooisms, such as having GNU coreutils!

`cp -a` is a non-portable extension, which fails on at least some macOS
systems (as reported on IRC for macosx ppc 10.4.11). It is documented by
GNU as being one of several shortcut options with the following
equivalencies:

`-a` ---> `-dR --preserve=all`
`-d` ---> `-P --preserve=links`
`-p` ---> `--preserve=mode,ownership,timestamps`
`--preserve=all` ---> `--preserve=mode,ownership,timestamps,links,context,xattr`

Hence `-a` could equally be `-RpP --preserve=links,context,xattr`

FreeBSD has `-a` too, and documents it as a shortcut for `-RpP`.

We use it twice, once to install a just-compiled executable and once to
bootstrap /usr/share/portage. We have no need for the extra GNU
--preserve's. Moreover, -RpP is all documented in POSIX 2008, so it is
sure to work everywhere -- and we use these options elsewhere in this
script already. ;)

Fixes: 636a653fcc42a974748a26cbf6c999592c762466
Acked-by: Sam James <sam@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>