scripts/bootstrap-prefix: call configure_cflags in stage1 too
While most packages historically were stand alone and not requiring
libraries, Python was handled as exception to this.  With the possible
bootstrap of a compiler, we need proper setup to find the temp
location's installed libraries and headers, so just call configure_cflags
for that purpose.

Ensure we build a bootstrap compiler (GCC-5) in stage1 if the host
compiler is not sufficient, do this after we build Python, so we have
libraries like zlib to take advantage of.  Because LDFLAGS is now set
with the proper values, the built xgcc can now find the stage1 built
libraries.

Finally, speed up the build on machines with less CPUs (as typically
found in VMs these days) by taking all if less than 4 are available.

Closes: https://bugs.gentoo.org/967262
Signed-off-by: Fabian Groffen <grobian@gentoo.org>