scripts/bootstrap-prefix: become quieter in stage2 and 3
Use --jobs=1 to trigger Portage's more quiet mode.  The amount of
logging is just swamping the terminal, and there's rarely a need for it,
given that on error Portage dumps the full output.

It is possible to also enable parallel mode this way, but there are some
dependency issues, and besides, there's a lot of single order required
here, so it doesn't pay off that much.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>